+++ src/lib/xiaomi.ts
@@ -4,13 +4,13 @@ import {
     calibrateAndPrecisionRoundOptionsIsPercentual,
     postfixWithEndpointName,
     precisionRound,
-    getKey,
     assertNumber,
     getFromLookup,
 } from './utils';
 
 import * as exposes from './exposes';
 import * as globalStore from './store';
+import * as constants from './constants';
 import {Fz, Definition, KeyValue, KeyValueAny} from './types';
 import * as modernExtend from './modernExtend';
 
@@ -514,11 +514,6 @@ export const numericAttributes2Payload = async (msg: Fz.Message, meta: Fz.Meta, 
                 payload.gas_sensitivity = getFromLookup(value, {1: '15%LEL', 2: '10%LEL'});
             }
             break;
-        case '276':
-            if (['VOCKQJK11LM'].includes(model.model)) {
-                payload.display_unit = getKey(VOCKQJK11LMDisplayUnit, value);
-            }
-            break;
         case '293':
             payload.click_mode = getFromLookup(value, {1: 'fast', 2: 'multi'});
             break;
@@ -766,6 +761,12 @@ export const numericAttributes2Payload = async (msg: Fz.Message, meta: Fz.Meta, 
         case 'illuminance':
             // It contains the illuminance and occupancy, but in z2m we use a custom timer to do it, so we ignore it
             break;
+        case 'displayUnit':
+            // Use aqaraDisplayUnit modernExtend, but we add it here to not shown an unknown key in the log
+            break;
+        case 'airQuality':
+            // Use aqaraAirQuality modernExtend, but we add it here to not shown an unknown key in the log
+            break;
         default:
             if (meta.logger) meta.logger.debug(`${model.model}: unknown key ${key} with value ${value}`);
         }
@@ -774,13 +775,6 @@ export const numericAttributes2Payload = async (msg: Fz.Message, meta: Fz.Meta, 
     if (meta.logger) meta.logger.debug(`${model.model}: Processed data into payload ${JSON.stringify(payload)}`);
 
     return payload;
-};
-
-export const VOCKQJK11LMDisplayUnit = {
-    'mgm3_celsius': 0x00, // mg/m³, °C (default)
-    'ppb_celsius': 0x01, // ppb, °C
-    'mgm3_fahrenheit': 0x10, // mg/m³, °F
-    'ppb_fahrenheit': 0x11, // ppb, °F
 };
 
 export const numericAttributes2Options = (definition: Definition) => {
@@ -1352,7 +1346,7 @@ export const xiaomiModernExtend = {
         name: 'switch_type',
         lookup: {'toggle': 1, 'momentary': 2, 'none': 3},
         cluster: 'aqaraOpple',
-        attribute: {id: 0x000a, type: 0x20},
+        attribute: {ID: 0x000a, type: 0x20},
         description: 'External switch type',
         zigbeeCommandOptions: {manufacturerCode},
         ...args,
@@ -1361,7 +1355,7 @@ export const xiaomiModernExtend = {
         name: 'power_on_behavior',
         lookup: {'on': 0, 'previous': 1, 'off': 2},
         cluster: 'aqaraOpple',
-        attribute: {id: 0x0517, type: 0x20},
+        attribute: {ID: 0x0517, type: 0x20},
         description: 'Controls the behavior when the device is powered on after power loss',
         zigbeeCommandOptions: {manufacturerCode},
         ...args,
@@ -1370,7 +1364,7 @@ export const xiaomiModernExtend = {
         name: 'operation_mode',
         lookup: {'decoupled': 0, 'control_relay': 1},
         cluster: 'aqaraOpple',
-        attribute: {id: 0x0200, type: 0x20},
+        attribute: {ID: 0x0200, type: 0x20},
         description: 'Decoupled mode for relay',
         zigbeeCommandOptions: {manufacturerCode},
         ...args,
@@ -1379,6 +1373,43 @@ export const xiaomiModernExtend = {
         lookup: {'single': 1},
         cluster: 'genMultistateInput',
         attribute: 'presentValue',
+        ...args,
+    }),
+    aqaraVoc: (args?: Partial<modernExtend.NumericArgs>) => modernExtend.numeric({
+        name: 'voc',
+        cluster: 'genAnalogInput',
+        attribute: 'presentValue',
+        configureReporting: {minimumReportInterval: 10, maximumReportInterval: constants.repInterval.HOUR, reportableChange: 5},
+        endpointID: 1,
+        description: 'Measured VOC value',
+        unit: 'ppb',
+        readOnly: true,
+        ...args,
+    }),
+    aqaraAirQuality: (args?: Partial<modernExtend.EnumLookupArgs>) => modernExtend.enumLookup({
+        name: 'air_quality',
+        lookup: {'excellent': 1, 'good': 2, 'moderate': 3, 'poor': 4, 'unhealthy': 5},
+        cluster: 'aqaraOpple',
+        attribute: 'airQuality',
+        endpointID: 1,
+        zigbeeCommandOptions: {disableDefaultResponse: true},
+        description: 'Measured air quality',
+        readOnly: true,
+        ...args,
+    }),
+    aqaraDisplayUnit: (args?: Partial<modernExtend.EnumLookupArgs>) => modernExtend.enumLookup({
+        name: 'display_unit',
+        lookup: {
+            'mgm3_celsius': 0x00, // mg/m³, °C (default)
+            'ppb_celsius': 0x01, // ppb, °C
+            'mgm3_fahrenheit': 0x10, // mg/m³, °F
+            'ppb_fahrenheit': 0x11, // ppb, °F
+        },
+        cluster: 'aqaraOpple',
+        attribute: 'displayUnit',
+        endpointID: 1,
+        zigbeeCommandOptions: {disableDefaultResponse: true},
+        description: 'Units to show on the display',
         ...args,
     }),
 };
@@ -1420,7 +1451,6 @@ export const fromZigbee = {
 exports.buffer2DataObject = buffer2DataObject;
 exports.numericAttributes2Payload = numericAttributes2Payload;
 exports.numericAttributes2Options = numericAttributes2Options;
-exports.VOCKQJK11LMDisplayUnit = VOCKQJK11LMDisplayUnit;
 exports.fp1 = fp1;
 exports.trv = trv;
 exports.manufacturerCode = manufacturerCode;